requirements.txt
pyserial >1.3.4

